+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Rename packagemanagersync to packagemanagerservicehtml
 * Move non update/download package methods from packagemanager to packagemanagerservicejson
 * Add pyaemresult#debug
+* Add packagemanagerservicejsp#is_package_uploaded
 
 ### 0.9.1
 * Add file_path arg to webconsole#install_bundle to allow custom bundle file location